+++ src/transaction/index.ts
@@ -1,2 +1,5 @@
 export * from './firestore-pubsub/index.js';
+export * from './spanner-outbox/index.js';
 export * from './spanner-pubsub/index.js';
+export { SpannerStateTransaction } from './spanner-state-transaction.js';
+export { SpannerTransaction } from './spanner-transaction.js';