+++ src/transaction/spanner-pubsub/runner.ts
@@ -1,30 +1,18 @@
-import {
-  BufferEventTransaction,
-  TransactionOldTimestampError,
-  TransactionRunner,
-} from '@causa/runtime';
+import { BufferEventTransaction, TransactionRunner } from '@causa/runtime';
 import { Logger } from '@causa/runtime/nestjs';
 import { Injectable } from '@nestjs/common';
-import { setTimeout } from 'timers/promises';
 import { PubSubPublisher } from '../../pubsub/index.js';
-import {
-  SpannerEntityManager,
-  TemporarySpannerError,
-} from '../../spanner/index.js';
-import { SpannerStateTransaction } from './state-transaction.js';
-import { SpannerPubSubTransaction } from './transaction.js';
-
-/**
- * The delay, in milliseconds, over which a timestamp issue is deemed irrecoverable.
- */
-const ACCEPTABLE_PAST_DATE_DELAY = 25000;
+import { SpannerEntityManager } from '../../spanner/index.js';
+import { SpannerStateTransaction } from '../spanner-state-transaction.js';
+import { SpannerTransaction } from '../spanner-transaction.js';
+import { throwRetryableInTransactionIfNeeded } from '../spanner-utils.js';
 
 /**
  * A {@link TransactionRunner} that uses Spanner for state and Pub/Sub for events.
  * A Spanner transaction is used as the main transaction. If it succeeds, events are published to Pub/Sub outside of it.
  */
 @Injectable()
-export class SpannerPubSubTransactionRunner extends TransactionRunner<SpannerPubSubTransaction> {
+export class SpannerPubSubTransactionRunner extends TransactionRunner<SpannerTransaction> {
   /**
    * Creates a new {@link SpannerPubSubTransactionRunner}.
    *
@@ -41,7 +29,7 @@ export class SpannerPubSubTransactionRunner extends TransactionRunner<SpannerPub
   }
 
   async run<T>(
-    runFn: (transaction: SpannerPubSubTransaction) => Promise<T>,
+    runFn: (transaction: SpannerTransaction) => Promise<T>,
   ): Promise<[T]> {
     this.logger.info('Creating a Spanner Pub/Sub transaction.');
 
@@ -53,7 +41,7 @@ export class SpannerPubSubTransactionRunner extends TransactionRunner<SpannerPub
         );
         // This must be inside the Spanner transaction because staged messages should be cleared when the transaction is retried.
         const eventTransaction = new BufferEventTransaction(this.publisher);
-        const transaction = new SpannerPubSubTransaction(
+        const transaction = new SpannerTransaction(
           stateTransaction,
           eventTransaction,
         );
@@ -64,23 +52,8 @@ export class SpannerPubSubTransactionRunner extends TransactionRunner<SpannerPub
           this.logger.info('Committing the Spanner transaction.');
           return { result, eventTransaction };
         } catch (error) {
-          // `TransactionOldTimestampError`s indicate that the transaction is using a timestamp older than what is
-          // observed in the state (Spanner).
-          // Throwing a `SpannerTransactionOldTimestampError` will cause the transaction to be retried with a newer
-          // timestamp.
-          if (!(error instanceof TransactionOldTimestampError)) {
-            throw error;
-          }
-
-          const delay = error.delay ?? Infinity;
-          if (delay >= ACCEPTABLE_PAST_DATE_DELAY) {
-            throw error;
-          }
-          if (delay > 0) {
-            await setTimeout(delay);
-          }
-
-          throw TemporarySpannerError.retryableInTransaction(error.message);
+          await throwRetryableInTransactionIfNeeded(error);
+          throw error;
         }
       },
     );